Unit tests for isers and groups management (admin section)
