+++ polls/templates/polls/adm_groups.html
@@ -29,7 +29,7 @@
                             {{ grp.group_name }} {{ grp.weight }}
 
                             <!-- Update buttons -->
-                            <a type="button" class="update-user btn btn-sm" href="{% url 'polls:adm_group_detail' company.comp_slug grp.id %}">
+                            <a type="button" class="update-grp btn btn-sm" href="{% url 'polls:adm_group_detail' company.comp_slug grp.id %}">
                                 <span class="fa fa-pencil" style="color: #339af0;" data-toggle="tooltip" title="Modifier"></span>
                             </a>
 
@@ -47,7 +47,7 @@
 
                 <div class="text-center">
                     <!-- Create button -->
-                    <a type="button" class="create-grp btn btn-primary" href="{% url 'polls:adm_group_detail' company.comp_slug %}">Nouveau groupe</a>
+                    <a type="button" class="create-grp btn btn-primary" href="{% url 'polls:adm_create_group' company.comp_slug %}">Nouveau groupe</a>
                 </div>
 
                 <div>
