+++ polls/tests_admin.py
@@ -38,7 +38,6 @@ from .forms import (
 )
 
 
-
 class TestOptions(TestCase):
     def setUp(self):
         self.company = create_dummy_company("Société de test")
@@ -98,7 +97,6 @@ class TestOptions(TestCase):
             comp_data
         )
         self.company.refresh_from_db()
-        # print(self.company.__dict__)
         self.assertEqual(response.status_code, 200)
         self.assertContains(response, "987654321")
         self.assertNotContains(response, "0123456789")
@@ -129,10 +127,18 @@ class TestAdmUsers(TestCase):
             if usr.is_admin: i += 1
         self.assertEqual(i, 2)
 
-    def test_delete_user(self):
-        url = reverse("polls:adm_delete_user", args=[self.user13.id])
-        response = self.client.get(url)
+    def test_adm_delete_user(self):
+        test_usercomp_id = self.usr13.id
+        usrcomp = UserComp.objects.get(user__username="user13")
+        self.assertEqual(usrcomp.id, test_usercomp_id)
 
+        url = reverse("polls:adm_delete_user", args=[self.company.comp_slug, test_usercomp_id])
+        response = self.client.get(url)
+        self.assertEqual(response.status_code, 302)
+        with self.assertRaises(User.DoesNotExist):
+            User.objects.get(id=test_usercomp_id)
+        with self.assertRaises(UserComp.DoesNotExist):
+            UserComp.objects.get(id=test_usercomp_id)
 
 
 class TestUserProfile(TestCase):
@@ -205,16 +211,144 @@ class TestUserProfile(TestCase):
         self.assertContains(response, "user11")
         self.assertContains(response, "user11@toto.com")
 
-        # Profil update
+        # Profile update by admin user
+        self.client.force_login(self.user_staff.user)
         response = self.client.post(
             reverse("polls:adm_user_profile", args=[self.company.comp_slug, self.usr11.user.id]),
             {"email": "new_mail@titi.com"}
         )
         self.usr11.refresh_from_db()
-        self.assertEqual(response.status_code, 200)
-        self.assertContains(response, "user11")
-        self.assertNotContains(response, "user11@toto.com")
-        self.assertContains(response, "new_mail@titi.com")
+        self.assertEqual(response.status_code, 302)
         self.assertEqual(self.usr11.user.email, "new_mail@titi.com")
 
 
+class TestAdmGroups(TestCase):
+    def setUp(self):
+        self.company = create_dummy_company("Société de test")
+
+        self.user_staff = create_dummy_user(self.company, "staff", admin=True)
+        self.usr11 = create_dummy_user(self.company, "user11")
+        self.usr12 = create_dummy_user(self.company, "user12", admin=True)
+        self.usr13 = create_dummy_user(self.company, "user13")
+        self.usr14 = create_dummy_user(self.company, "user14")
+        self.usr21 = create_dummy_user(self.company, "user21")
+        self.usr22 = create_dummy_user(self.company, "user22")
+
+        user_list = [self.usr11.id, self.usr12.id, self.usr13.id, self.usr14.id]
+        users = UserComp.objects.filter(id__in=user_list)
+        self.group1 = EventGroup.create_group({
+            "company": self.company,
+            "group_name": "Groupe 1",
+            "weight": 40,
+            },
+            user_list=users)
+
+        user_list = [self.usr21.id, self.usr22.id]
+        users = UserComp.objects.filter(id__in=user_list)
+        self.group2 = EventGroup.create_group({
+            "company": self.company,
+            "group_name": "Groupe 2",
+            "weight": 60,
+            },
+            user_list=users)
+
+    def test_adm_groups(self):
+        # Test access with no connection
+        url = reverse("polls:adm_groups", args=[self.company.comp_slug])
+        response = self.client.get(url)
+        self.assertEqual(response.status_code, 302)
+
+        # Test access for a dummy user
+        self.client.force_login(self.usr11.user)
+        url = reverse("polls:adm_groups", args=[self.company.comp_slug])
+        response = self.client.get(url)
+        self.assertEqual(response.status_code, 302)
+
+        # Test access for a staff user
+        self.client.force_login(self.user_staff.user)
+        url = reverse("polls:adm_groups", args=[self.company.comp_slug])
+        response = self.client.get(url)
+        self.assertEqual(response.status_code, 200)
+        self.assertEqual(len(response.context['group_list']), 2)
+
+    def test_adm_create_group(self):
+        self.client.force_login(self.user_staff.user)
+
+        user_list = [self.usr11.id, self.usr21.id]
+        users = "-".join([str(x) for x in user_list])
+        group_form = {}
+        group_form["all_users"] = UserComp.objects.all()
+        group_form["users"] = UserComp.objects.filter(id__in=user_list)
+
+        response = self.client.post(
+            reverse("polls:adm_create_group", args=[self.company.comp_slug]),
+            {"company": self.company,
+            "group_name": "Nouveau Groupe",
+            "weight": 12,
+            "group_form": group_form,
+            "users_in_group": users}
+        )
+        self.assertEqual(response.status_code, 200)
+        new_group = EventGroup.objects.get(group_name="Nouveau Groupe")
+        self.assertEqual(new_group.id, 3)
+        group_users = UserComp.objects.filter(eventgroup__id=new_group.id)
+        self.assertEqual(len(group_users), 2)
+        self.assertIn(self.usr11, group_users)
+        self.assertEqual(group_users[0].id, self.usr11.id)
+
+        # Create group without any user
+        response = self.client.post(
+            reverse("polls:adm_create_group", args=[self.company.comp_slug]),
+            {"company": self.company,
+            "group_name": "Groupe Vide",
+            "weight": 33,
+            "group_form": group_form,
+            "users_in_group": []}
+        )
+        self.assertEqual(response.status_code, 200)
+        new_group = EventGroup.objects.get(group_name="Groupe Vide")
+        group_users = UserComp.objects.filter(eventgroup__id=new_group.id)
+        self.assertEqual(len(group_users), 0)
+
+
+    def test_adm_update_group(self):
+        self.client.force_login(self.user_staff.user)
+        group_users = self.group1.users.all()
+        self.assertIn(self.usr11, group_users)
+        self.assertIn(self.usr14, group_users)
+        self.assertNotIn(self.usr21, group_users)
+
+        user_list = [self.usr11.id, self.usr21.id]
+        users = "-".join([str(x) for x in user_list])
+        group_form = {}
+        group_form["all_users"] = UserComp.objects.all()
+        group_form["users"] = UserComp.objects.filter(id__in=user_list)
+
+        response = self.client.post(
+            reverse("polls:adm_group_detail", args=[self.company.comp_slug, self.group1.id]),
+            {"company": self.company,
+            "group_name": "Groupe 1",
+            "weight": 33,
+            "group_form": group_form,
+            "users_in_group": users}
+        )
+
+        self.assertEqual(response.status_code, 200)
+        self.group1.refresh_from_db()
+        self.assertEqual(self.group1.weight, 33)
+        group_users = response.context["group_form"].fields["users"].queryset
+        self.assertIn(self.usr11, group_users)
+        self.assertNotIn(self.usr14, group_users)
+        self.assertIn(self.usr21, group_users)
+
+
+    def test_adm_delete_group(self):
+        test_group_id = self.group1.id
+        grp = EventGroup.objects.get(group_name="Groupe 1")
+        self.assertEqual(grp.id, test_group_id)
+
+        url = reverse("polls:adm_delete_group", args=[self.company.comp_slug, test_group_id])
+        response = self.client.get(url)
+        self.assertEqual(response.status_code, 302)
+        with self.assertRaises(EventGroup.DoesNotExist):
+            EventGroup.objects.get(id=test_group_id)
